update router, add 404 page, dashboard and logout, save user token in localStorage, move cors middleware
