+++ internal/http/router.go
@@ -6,7 +6,6 @@ import (
 	"github.com/DevPulseLab/salat/internal/config"
 	"github.com/DevPulseLab/salat/internal/db/models"
 	"github.com/DevPulseLab/salat/internal/handlers"
-	"github.com/DevPulseLab/salat/internal/http/middleware"
 	"github.com/DevPulseLab/salat/internal/middlewares"
 	"github.com/gin-gonic/gin"
 	"gorm.io/gorm"
@@ -18,7 +17,7 @@ func InitializeRoutes(router *gin.Engine, db *gorm.DB, config *config.Config) {
 	authHandler := handlers.NewAuthHandler(db, config)
 	userHandler := handlers.NewUserHandler(db)
 
-	router.Use(middleware.CORSMiddleware())
+	router.Use(middlewares.CORSMiddleware())
 
 	router.GET("/api/ping", func(ctx *gin.Context) {
 		ctx.JSON(http.StatusOK, map[string]string{"ping": "pong"})
