+++ internal/middlewares/cors_middleware.go
@@ -1,4 +1,4 @@
-package middleware
+package middlewares
 
 import "github.com/gin-gonic/gin"
